refactor(kv-pool): consolidate layerwise cache layout planning
Build the complete layerwise reuse layout through one domain-specific entry point, remove thin configuration helpers, and keep load-prefix policy at its runtime call site.
